+++ packages/frontend/component/src/components/affine-banner/local-demo-tips.tsx
@@ -1,46 +1,24 @@
-import { Button, IconButton } from '@affine/component/ui/button';
-import { useI18n } from '@affine/i18n';
+import { IconButton } from '@affine/component/ui/button';
 import { CloseIcon } from '@blocksuite/icons/rc';
-import { cssVar } from '@toeverything/theme';
-import { useCallback } from 'react';
 
 import * as styles from './index.css';
 
 type LocalDemoTipsProps = {
-  isLoggedIn: boolean;
-  onLogin: () => void;
-  onEnableCloud: () => void;
+  isLoggedIn?: boolean;
+  onLogin?: () => void;
+  onEnableCloud?: () => void;
   onClose: () => void;
 };
 
-export const LocalDemoTips = ({
-  onClose,
-  isLoggedIn,
-  onLogin,
-  onEnableCloud,
-}: LocalDemoTipsProps) => {
-  const t = useI18n();
-  const buttonLabel = isLoggedIn
-    ? t['Enable AFFiNE Cloud']()
-    : t['Sign in and Enable']();
-
-  const handleClick = useCallback(() => {
-    if (isLoggedIn) {
-      return onEnableCloud();
-    }
-    return onLogin();
-  }, [isLoggedIn, onEnableCloud, onLogin]);
-
+export const LocalDemoTips = ({ onClose }: LocalDemoTipsProps) => {
   return (
     <div className={styles.tipsContainer} data-testid="local-demo-tips">
       <div className={styles.tipsMessage}>
-        {t['com.affine.banner.local-warning']()}
+        Your data is stored locally on this device. Export regularly to keep it
+        safe.
       </div>
 
       <div className={styles.tipsRightItem}>
-        <Button style={{ background: cssVar('white') }} onClick={handleClick}>
-          {buttonLabel}
-        </Button>
         <IconButton
           onClick={onClose}
           size="20"