+++ packages/frontend/core/src/components/root-app-sidebar/user-info/account-menu.tsx
@@ -1,20 +1,16 @@
 import { MenuItem } from '@affine/component';
-import { ServerService, UserFeatureService } from '@affine/core/modules/cloud';
 import { WorkspaceDialogService } from '@affine/core/modules/dialogs';
 import { useI18n } from '@affine/i18n';
 import { track } from '@affine/track';
-import { AccountIcon, AdminIcon, SignOutIcon } from '@blocksuite/icons/rc';
-import { useLiveData, useService } from '@toeverything/infra';
-import { useCallback, useEffect } from 'react';
+import { AccountIcon, SignOutIcon } from '@blocksuite/icons/rc';
+import { useService } from '@toeverything/infra';
+import { useCallback } from 'react';
 
 import { useSignOut } from '../../hooks/affine/use-sign-out';
 
 export const AccountMenu = () => {
   const workspaceDialogService = useService(WorkspaceDialogService);
   const openSignOutModal = useSignOut();
-  const serverService = useService(ServerService);
-  const userFeatureService = useService(UserFeatureService);
-  const isAFFiNEAdmin = useLiveData(userFeatureService.userFeature.isAdmin$);
 
   const onOpenAccountSetting = useCallback(() => {
     track.$.navigationPanel.profileAndBadge.openSettings({ to: 'account' });
@@ -23,15 +19,7 @@ export const AccountMenu = () => {
     });
   }, [workspaceDialogService]);
 
-  const onOpenAdminPanel = useCallback(() => {
-    window.open(`${serverService.server.baseUrl}/admin`, '_blank');
-  }, [serverService.server.baseUrl]);
-
   const t = useI18n();
-
-  useEffect(() => {
-    userFeatureService.userFeature.revalidate();
-  }, [userFeatureService]);
 
   return (
     <>
@@ -42,15 +30,6 @@ export const AccountMenu = () => {
       >
         {t['com.affine.workspace.cloud.account.settings']()}
       </MenuItem>
-      {isAFFiNEAdmin ? (
-        <MenuItem
-          prefixIcon={<AdminIcon />}
-          data-testid="workspace-modal-account-admin-option"
-          onClick={onOpenAdminPanel}
-        >
-          {t['com.affine.workspace.cloud.account.admin']()}
-        </MenuItem>
-      ) : null}
       <MenuItem
         prefixIcon={<SignOutIcon />}
         data-testid="workspace-modal-sign-out-option"