+++ packages/frontend/core/src/components/top-tip.tsx
@@ -1,19 +1,13 @@
 import { BrowserWarning, LocalDemoTips } from '@affine/component/affine-banner';
 import { Trans, useI18n } from '@affine/i18n';
-import { useLiveData, useService } from '@toeverything/infra';
-import { useCallback, useState } from 'react';
+import { useState } from 'react';
 
-import { useEnableCloud } from '../components/hooks/affine/use-enable-cloud';
-import { AuthService } from '../modules/cloud';
-import { GlobalDialogService } from '../modules/dialogs';
 import type { Workspace } from '../modules/workspace';
 
 const minimumChromeVersion = 106;
 
 const shouldShowWarning = (() => {
   if (BUILD_CONFIG.isElectron) {
-    // even though desktop has compatibility issues,
-    //  we don't want to show the warning
     return false;
   }
   if (BUILD_CONFIG.isMobileEdition) {
@@ -33,7 +27,6 @@ const OSWarningMessage = () => {
     environment.chromeVersion &&
     environment.chromeVersion < minimumChromeVersion;
 
-  // TODO(@L-Sun): remove this message when mobile version is able to edit.
   if (environment.isMobile) {
     return <span>{t['com.affine.top-tip.mobile']()}</span>;
   }
@@ -55,23 +48,13 @@ const OSWarningMessage = () => {
 };
 
 export const TopTip = ({
-  pageId,
   workspace,
 }: {
   pageId?: string;
   workspace: Workspace;
 }) => {
-  const loginStatus = useLiveData(useService(AuthService).session.status$);
-  const isLoggedIn = loginStatus === 'authenticated';
-
   const [showWarning, setShowWarning] = useState(shouldShowWarning);
   const [showLocalDemoTips, setShowLocalDemoTips] = useState(true);
-  const confirmEnableCloud = useEnableCloud();
-
-  const globalDialogService = useService(GlobalDialogService);
-  const onLogin = useCallback(() => {
-    globalDialogService.open('sign-in', {});
-  }, [globalDialogService]);
 
   if (
     !BUILD_CONFIG.isElectron &&
@@ -80,11 +63,6 @@ export const TopTip = ({
   ) {
     return (
       <LocalDemoTips
-        isLoggedIn={isLoggedIn}
-        onLogin={onLogin}
-        onEnableCloud={() =>
-          confirmEnableCloud(workspace, { openPageId: pageId })
-        }
         onClose={() => {
           setShowLocalDemoTips(false);
         }}